Refactored code after mentor review.
