+++ src/main/java/com/loievroman/bookstoreapp/exception/CustomGlobalExceptionHandler.java
@@ -1,6 +1,5 @@
 package com.loievroman.bookstoreapp.exception;
 
-import io.jsonwebtoken.ExpiredJwtException;
 import java.time.LocalDateTime;
 import java.util.LinkedHashMap;
 import java.util.List;
@@ -12,7 +11,6 @@ import org.springframework.validation.ObjectError;
 import org.springframework.web.bind.MethodArgumentNotValidException;
 import org.springframework.web.bind.annotation.ControllerAdvice;
 import org.springframework.web.bind.annotation.ExceptionHandler;
-import org.springframework.web.bind.annotation.ResponseBody;
 
 @ControllerAdvice
 public class CustomGlobalExceptionHandler {
@@ -59,12 +57,6 @@ public class CustomGlobalExceptionHandler {
                 ex.getMessage(),
                 HttpStatus.BAD_REQUEST
         );
-    }
-
-    @ExceptionHandler(ExpiredJwtException.class)
-    @ResponseBody
-    public ResponseEntity<String> handleExpiredJwtException(ExpiredJwtException ex) {
-        return new ResponseEntity<>("JWT has expired.", HttpStatus.UNAUTHORIZED);
     }
 
     private String getErrorMessage(ObjectError objectError) {
